+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 
 * Added `kubernetes.network.apiVIP6` field to enable cluster LoadBalancer based on IPv6 address
 * Added the `enableExtras` flag to enable the SUSE Linux Extras repository during RPM resolution.
+* Added support for custom registration server (like RMT/SUMA) via `sccRegistrationUrl`.
 
 ### Image Configuration Directory Changes